refactor(event-trigger-jsonschema): change project name
