+++ event-trigger-jsonschema/build.gradle.kts
@@ -1,7 +1,6 @@
 dependencies {
     api(project(":schedulerx"))
     api(VertxLibs.jsonSchema)
-
     compileOnly(UtilLibs.jetbrainsAnnotations)
 
     testImplementation(testFixtures(project(":schedulerx")))
